+++ apps/e2e/tests/backend/endpoints/api/v1/auth/password/send-reset-code.test.ts
@@ -25,13 +25,13 @@ it("should send a password reset code per e-mail", async ({ expect }) => {
       MailboxMessage {
         "from": "Stack Dashboard <noreply@example.com>",
         "subject": "Verify your email at Stack Dashboard",
-        "to": ["<<stripped UUID>@stack-generated.example.com>"],
+        "to": ["<default-mailbox--<stripped UUID>@stack-generated.example.com>"],
         <some fields may have been hidden>,
       },
       MailboxMessage {
         "from": "Stack Dashboard <noreply@example.com>",
         "subject": "Reset your password at Stack Dashboard",
-        "to": ["<<stripped UUID>@stack-generated.example.com>"],
+        "to": ["<default-mailbox--<stripped UUID>@stack-generated.example.com>"],
         <some fields may have been hidden>,
       },
     ]