+++ apps/e2e/tests/backend/endpoints/api/v1/contact-channels/send-verification-code.test.ts
@@ -40,13 +40,13 @@ it("should send a verification code per e-mail", async ({ expect }) => {
       MailboxMessage {
         "from": "Stack Dashboard <noreply@example.com>",
         "subject": "Verify your email at Stack Dashboard",
-        "to": ["<<stripped UUID>@stack-generated.example.com>"],
+        "to": ["<default-mailbox--<stripped UUID>@stack-generated.example.com>"],
         <some fields may have been hidden>,
       },
       MailboxMessage {
         "from": "Stack Dashboard <noreply@example.com>",
         "subject": "Verify your email at Stack Dashboard",
-        "to": ["<<stripped UUID>@stack-generated.example.com>"],
+        "to": ["<default-mailbox--<stripped UUID>@stack-generated.example.com>"],
         <some fields may have been hidden>,
       },
     ]