+++ packages/stack/src/lib/stack-app.ts
@@ -26,7 +26,7 @@ import { deepPlainEquals, filterUndefined, omit } from "@stackframe/stack-shared
 import { ReactPromise, neverResolve, runAsynchronously, wait } from "@stackframe/stack-shared/dist/utils/promises";
 import { suspend, suspendIfSsr } from "@stackframe/stack-shared/dist/utils/react";
 import { Result } from "@stackframe/stack-shared/dist/utils/results";
-import { Store } from "@stackframe/stack-shared/dist/utils/stores";
+import { Store, storeLock } from "@stackframe/stack-shared/dist/utils/stores";
 import { mergeScopeStrings } from "@stackframe/stack-shared/dist/utils/strings";
 import { getRelativePart, isRelative } from "@stackframe/stack-shared/dist/utils/urls";
 import { generateUuid } from "@stackframe/stack-shared/dist/utils/uuids";
@@ -251,6 +251,7 @@ function useStore<T>(store: Store<T>): T {
   return React.useSyncExternalStore(subscribe, getSnapshot, getSnapshot);
 }
 
+/** @internal */
 export const stackAppInternalsSymbol = Symbol.for("StackAuth--DO-NOT-USE-OR-YOU-WILL-BE-FIRED--StackAppInternals");
 
 const allClientApps = new Map<string, [checkString: string, app: StackClientApp<any, any>]>();
@@ -1536,12 +1537,14 @@ class _StackClientAppImpl<HasTokenStore extends boolean, ProjectId extends strin
   }
 
   protected async _signOut(session: InternalSession, options?: { redirectUrl?: URL | string }): Promise<void> {
-    await this._interface.signOut(session);
-    if (options?.redirectUrl) {
-      await _redirectTo(options.redirectUrl);
-    } else {
-      await this.redirectToAfterSignOut();
-    }
+    await storeLock.withWriteLock(async () => {
+      await this._interface.signOut(session);
+      if (options?.redirectUrl) {
+        await _redirectTo(options.redirectUrl);
+      } else {
+        await this.redirectToAfterSignOut();
+      }
+    });
   }
 
   async signOut(options?: { redirectUrl?: URL | string }): Promise<void> {
@@ -2266,6 +2269,9 @@ class _StackAdminAppImpl<HasTokenStore extends boolean, ProjectId extends string
   private readonly _svixTokenCache = createCache(async () => {
     return await this._interface.getSvixToken();
   });
+  private readonly _metricsCache = createCache(async () => {
+    return await this._interface.getMetrics();
+  });
 
   constructor(options: StackAdminAppConstructorOptions<HasTokenStore, ProjectId>) {
     super({
@@ -2517,6 +2523,35 @@ class _StackAdminAppImpl<HasTokenStore extends boolean, ProjectId extends string
 
   protected async _refreshApiKeys() {
     await this._apiKeysCache.refresh([]);
+  }
+
+  get [stackAppInternalsSymbol]() {
+    return {
+      ...super[stackAppInternalsSymbol],
+      useMetrics: (): any => {
+        return useAsyncCache(this._metricsCache, [], "useMetrics()");
+      }
+    };
+  }
+
+  async sendTestEmail(options: {
+    recipientEmail: string,
+    emailConfig: EmailConfig,
+  }): Promise<Result<undefined, { errorMessage: string }>> {
+    const response = await this._interface.sendTestEmail({
+      recipient_email: options.recipientEmail,
+      email_config: {
+        ...options.emailConfig,
+        sender_email: options.emailConfig.senderEmail,
+        sender_name: options.emailConfig.senderName,
+      },
+    });
+
+    if (response.success) {
+      return Result.ok(undefined);
+    } else {
+      return Result.error({ errorMessage: response.error_message ?? throwErr("Email test error not specified") });
+    }
   }
 }
 
@@ -3410,6 +3445,11 @@ export type StackAdminApp<HasTokenStore extends boolean = boolean, ProjectId ext
     deleteTeamPermissionDefinition(permissionId: string): Promise<void>,
 
     useSvixToken(): string,
+
+    sendTestEmail(options: {
+      recipientEmail: string,
+      emailConfig: EmailConfig,
+    }): Promise<Result<undefined, { errorMessage: string }>>,
   }
   & StackServerApp<HasTokenStore, ProjectId>
 );
@@ -3453,8 +3493,11 @@ type AsyncStoreProperty<Name extends string, Args extends any[], Value, IsMultip
   & { [key in `${IsMultiple extends true ? "list" : "get"}${Capitalize<Name>}`]: (...args: Args) => Promise<Value> }
   & { [key in `use${Capitalize<Name>}`]: (...args: Args) => Value }
 
-type Prettify<T> = {
-  [K in keyof T]: T[K];
-} & {};
-
-type x = Prettify<CurrentUser>
+type EmailConfig = {
+  host: string,
+  port: number,
+  username: string,
+  password: string,
+  senderEmail: string,
+  senderName: string,
+}