+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/stack-dashboard
 
+## 2.7.0
+
+### Minor Changes
+
+- Various changes
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.0
+  - @stackframe/stack-ui@2.7.0
+  - @stackframe/stack@2.7.0
+  - @stackframe/stack-emails@2.7.0
+
 ## 2.6.39
 
 ### Patch Changes