+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @stackframe/e2e-tests
 
+## 2.7.0
+
+### Minor Changes
+
+- Various changes
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.0
+
 ## 2.6.39
 
 ### Patch Changes